+++ CHANGELOG.md
@@ -9,6 +9,10 @@
 - [TD-7091] Add language selector display messages for concepts
 - [TD-7049] Add AI Gemini messages
 
+### Changed
+
+- [TD-7119] Cache locales in controller
+
 ## [7.1.1] 2025-02-10
 
 ### Fixed